Extract individual type definitions from WireType
WireType itself is a recusrive type definition, because this is the
easiest to obtain from custom types using GHC.Generics. For code
generation and type difference calculation though we'll need
representations of the individual custom types. We now have a helper
function allowing us to get this representation.

We could wire this logic into the generics code, but mixing the logic
for obtaining wire types from generics representations with the logic
for flattening generics representations would complicate both.