Distinguish between custom and record types
A Haskell record type is only allowed to have a single constructor.
Otherwise the record field accessors will become partial functions.

A Haskell type intended to have multiple constructors can only allow
record type as constructor parameters, or it won't be able to change
parameters in a backwards compatible fashion.

This change makes it so we support both these types.