Support encoding and decoding Haskell sets
We encode and decode these into lists though. If we wouldn't do this,
we'd need to account for constraints on set items in other languages
being different than those in Haskell. For example, Haskell allows sets
to contain any element with an `Ord` instance. Elm's support for set
elements is much more constrained. I don't (at leats not at this point,
but likely never) want to write logic that would use a Set type on the
Elm side if the element type on the Elm side supports it.