Explicitly enable language extensions in modules
The default language extensions aren't used in compile-time error tests.
Listing extensions explicitly makes it possible to use more modules for
the compile time error tests.

The one extension that remains in package.yaml is StrictData. This
because it's easy to forget in modules, and not having it for compile
time tests should not result in different compiler errors.