Allow multiple constructors with records
I previously disallowed these because they result in partial functions.
GHC creates functions for accessing these fields, but if we pass a value
of the type created with a constructor that does not have the field that
results in a runtime error.

GHC supports the 'warn-partial-fields' flag to make you aware of this,
but I thought the custom type errors of this library might be nicer.

The thing is, the upcoming GHC 9.2 release will push out some changes
that will make records in Haskell much nicer to work with. Among them is
the NoFieldSelectors extension, which will prevent accessor functions
from being created. That will mean no partial functions can be created
either, rehabilitating the use of record notation in multi-constructor
types.